Small-scale experiment of a function transforming a “Parsimonious syntactic tree” into a “DuraLex semantic tree”
I guess such a method could work on larger syntactic trees, parametrized by
the property 'dynamic': here a quote is always after the article_id hence
article_id instruct the visitor to promote the current node to a parent. Lists
could be similarly instructed: when some rule is a list, the parent is put on
a stack and the current node is incremented always with the same parent (as
it is currently done in parse_definition_list).

With such method, I guess we could define the grammar on a side and the
translation table on the other side and it will create the same DuraLex
semantic tree (which will be then manipulated by its semantic visitors).