Improved ToSemanticTreeVisitor to be used top-down instead of bottom-up
Now rules passed as parameter to ToSemanticTreeVisitor can be entry rules,
for example the first Parsimious rule to create some (even empty) DuraLex node,
then child Parsimonious nodes will be treated as child DuraLex nodes.

Internally, it is followed the internal Parsimonious walk (i.e. bottom-up), but
newly-created child DuraLex nodes are grouped (or merged for non-typed DuraLex
nodes, which correspond to property definition of parent node) in the parent
DuraLex node (corresponding to a Parsimonious node). Some care is taken for
multiple children, even when the parent Parsimonious node is not a DuraLex
node. At the end we obtain a detached DuraLex tree, which must be attached to
the main DuraLex tree.