Fix issue with nodes only setting properties
I suspected this issue in the visitor 'Parsimonious tree to DuraLex tree'.
Properties are now separately handled and accumulated until a parent DuraLex
node use them. Consequently the recursion is now a tuple (DuraLex tree,
properties to be set in DuraLex parent).

For an easy usage, it is added a method in the visitor to attach the
converted-to-DuraLex Parsimonious tree on a DuraLex node.