WIP about hierarchical items in lists
Currently sentences like "Au deuxième alinéa, à la troisième phrase du
quatrième alinéa…" are recognised with legacy DuraLex as:
{"type": "alinea-reference", "order": 2, children: [
  {"type": "sentence-reference", "order": 3, children: [
    {"type": alinea-reference", "order": 4}]}]}
And would be with current ToSemanticTreeVisitor as:
{"type": "alinea-reference", "order": 2},
{"children": [
  {"type": "sentence-reference", "order": 3},
  {"type": "alinea-reference", "order": 4}]}
With this code it is possible to obtain (with a specific property for "du"):
{"type": "alinea-reference", "order": 2},
{"type": "alinea-reference", "order": 4, "children": [
  {"children": [{"type": "sentence-reference", "order": 3}]}]}
NB: an untyped node is currently created; it shouldn’t, and it triggers an
exception in a later DuraLex visitor.

The information carried by "du" should trigger some hierarchy.
I’m not completely sure about how to design this issue: other ideas
could be:
1/ to embed the hierarchy sentence < alinea < article directly in the visitor
   in order to decide if the children should form a list and/or a tree
2/ create trees by default and lists by exception.

PS: such mechanism, whatever it is exactly, is necessary for article 11 of
pjl XV-911 <http://www.assemblee-nationale.fr/15/textes/0911.asp#D_Article_11>.