Add more checks in RemoveQuotePrefixVisitor
In particular, check than the removed prefix "Art. {articleId}. -"
is the same than its container article-definition (if this one declared
the article name) and check that the dash in the articleId is the standard
dash (possibly the article-definition nodes could force their id to be a
standard dash).

Also the very last dash in "Art. {articleId}. -" can be any type of
dash between the 7 dashed defined in Unicode.

Also removed the multiline flag, because if there are multiple articles
defined in a single 'word' node, they should be splitted into independant
'article-definition'+'word' nodes (see #10).