Soften the rule about quote : most amendments from the Sénat don’t have quotes
In the Sénat, most amendments don’t have quotes, like this one:
```
Supprimer les mots :
: un enjeu social et sociétal majeur
```

The rule is now: search the first capturing rule between:
1/ terminate the current line, whatever there is on, pass whatever number of
   newlines, character quote, whatever newline beginning with a quote, ending
   quote; if recognised, the quote is between the two quote characters;
2/ pass whatever number of spaces, pass an optional colon, pass a newline,
   then the quote is the next line (only one line).

In the currently-recognised amendments it will change nothing, but in addition
more amendments from the Sénat will be recognised. A downside of this rule
is that unrecognised amendments could be even less recognised because one
line could be falsely recognised as a quote. But unrecognised amendments should
be recognised on the long term, hence it is not really an issue.