Syntax «» for quotes, distinct rules for singular/plural alineas, added tests
Previously only quotes with "" were recognised, now also «», particularly
given that they are the default quotes used in the amendments, this will
avoid a pre-processing step.

Now there are distinct rules for one or multiple alineas, I tend to prefer
two cases to avoid false positives, although this could be discussed and
judged against reality when DuraLex will be tested against a lot of real
amendments.

Finally given I removed the function parse_alinea_definition, it should
be renamed in the tests (into parse_definition). Also, added new tests
for alinea-definition, converted u'' to '' given we are Python3 now, and
added possibility to execute ParseAlineaDefinition as a main program.