Reconciliate a part of the tests with current code
About the recent change with lists. We are now with 13 failing tests
instead of 23.

Also, added a property 'count' with value 1 when it is known that there
is one alinea or one sentence.

Also converted some u'' to '' in tests given we are Python3; and
reordered some of the properties in tests, I find a canonical order
'type', then other properties then 'children' is easier to read.

Issue: #9