Reorganise links between parse_header* and parse_edit about references management
Historically references were searched at the beginning of parse_edit (it is
still the case for backward compatibility purposes) because most verbs had
the passive voice ("est remplacé", "est supprimé", …) and until some days there
was only one infinitive ("supprimer") as a special case. Recently was added
more infinitives ("remplacer", "ajouter", …).

I just worked on this amendment 15 of Sénat’s ppl 27 (2018):
  Alinéa 3, première phrase
  1° Supprimer les mots :
  à sa caisse d’assurance maladie d’affiliation
  2° Après les mots :
  au sens
  insérer les mots :
  et selon les modalités
Only the first and the two last lines were read, other were skipped (which was
worrying). In fact, we entered in parse_edit, found some references (hence we
didn’t leave with the first condition), we skipped almost all tokens until a
recognised token (“insérer”, because “supprimer” was not in the list because
it was a special case), we entered in the case for “insérer” and the end was
correctly handled.

To avoid such misinterpretation:
* “supprimer” is no more a special case, and it expects a reference after it;
* “est”/“sont” is mandatorily checked before each passive form (perhaps not
  necessary but I’m afraid of other edge-edge cases);
* we still search a reference at the beginning of parse_edit mainly for backward
  compatibility to keep immediate compatibility with working amendments, and if
  a couple "reference then verb" is not found we leave parse_edit without taking
  into account any partly-recognised reference;
* in higher levels (parse_header*), if parse_edit is not successful (we stayed
  at the same token) then we retry by searching a reference then possibly a
  verb (to be checked, perhaps we cannot logically find a verb), hence we have
  possibly found a reference, and then we search the parse_header of the lower
  level.
The resulting behaviour is that, in the case of a header with only a reference
but without any verb, a reference is added at the top of the children. This
reference is currently not used in further treatments but should be.

Currently this amendment is still not properly rendered by SedLex, but at least
the DuraLex tree looks much better.

I checked that the major forms of amendments are still recognised (new article,
deleted article, replace some words).